image-processor/requirements.txt
PyXDG==0.28
pillow==10.1.0
pillow_heif==0.14.0
